+++ pypsa/networks.py
@@ -12,7 +12,9 @@ import warnings
 from typing import TYPE_CHECKING, Any
 from weakref import ref
 
-from pypsa.common import deprecated_in_next_major, equals
+from deprecation import deprecated
+
+from pypsa.common import equals
 from pypsa.components.components import Components
 from pypsa.constants import DEFAULT_EPSG, DEFAULT_TIMESTAMP
 from pypsa.statistics.abstract import AbstractStatisticsAccessor
@@ -1346,7 +1348,9 @@ class Network(
         # Reindex to include all branches (even those not in cycles)
         return cycles_df.reindex(branches_i, fill_value=0).rename_axis(columns="cycle")
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `n.components.<component>` instead.",
     )
     def component(self, c_name: str) -> Component:
@@ -1359,7 +1363,11 @@ class Network(
         """
         return self.components[c_name]
 
-    @deprecated_in_next_major(details="Use `for component in n.components` instead.")
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
+        details="Use `for component in n.components` instead.",
+    )
     def iterate_components(
         self, components: Collection[str] | None = None, skip_empty: bool = True
     ) -> Iterator[Component]:
@@ -1625,7 +1633,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         branches = self.n.passive_branches()
         return branches[branches.sub_network == self.name]
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.<c_name>` instead.",
     )
     def component(self, c_name: str) -> SubNetworkComponents:
@@ -1642,7 +1652,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.components[c_name]
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.<c_name>.static` instead.",
     )
     def df(self, c_name: str) -> pd.DataFrame:
@@ -1659,7 +1671,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c[c_name].static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.<c_name>.static` instead.",
     )
     def static(self, c_name: str) -> pd.DataFrame:
@@ -1676,7 +1690,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.components[c_name].static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.<c_name>.dynamic` instead.",
     )
     def pnl(self, c_name: str) -> Dict:
@@ -1693,7 +1709,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c[c_name].dynamic
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.<c_name>.dynamic` instead.",
     )
     def dynamic(self, c_name: str) -> Dict:
@@ -1710,7 +1728,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.components[c_name].dynamic
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.buses.static.index` instead.",
     )
     def buses_i(self) -> pd.Index:
@@ -1727,7 +1747,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.buses.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.lines.static.index` instead.",
     )
     def lines_i(self) -> pd.Index:
@@ -1744,7 +1766,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.lines.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.transformers.static.index` instead.",
     )
     def transformers_i(self) -> pd.Index:
@@ -1761,7 +1785,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.transformers.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.generators.static.index` instead.",
     )
     def generators_i(self) -> pd.Index:
@@ -1778,7 +1804,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.generators.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.loads.static.index` instead.",
     )
     def loads_i(self) -> pd.Index:
@@ -1795,7 +1823,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.loads.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.shunt_impedances.static.index` instead.",
     )
     def shunt_impedances_i(self) -> pd.Index:
@@ -1812,7 +1842,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.shunt_impedances.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.storage_units.static.index` instead.",
     )
     def storage_units_i(self) -> pd.Index:
@@ -1829,7 +1861,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.storage_units.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.stores.index.static` instead.",
     )
     def stores_i(self) -> pd.Index:
@@ -1846,7 +1880,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.stores.static.index
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.buses.static` instead.",
     )
     def buses(self) -> pd.DataFrame:
@@ -1863,7 +1899,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.buses.static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.generators.static` instead.",
     )
     def generators(self) -> pd.DataFrame:
@@ -1880,7 +1918,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.generators.static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.loads.static` instead.",
     )
     def loads(self) -> pd.DataFrame:
@@ -1897,7 +1937,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.loads.static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.shunt_impedances.static` instead.",
     )
     def shunt_impedances(self) -> pd.DataFrame:
@@ -1914,7 +1956,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.shunt_impedances.static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `sub_network.components.storage_units.static` instead.",
     )
     def storage_units(self) -> pd.DataFrame:
@@ -1931,7 +1975,9 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.storage_units.static
 
-    @deprecated_in_next_major(
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
         details="Use `!!! deprecated.components.stores.static` instead.",
     )
     def stores(self) -> pd.DataFrame:
@@ -1943,7 +1989,11 @@ class SubNetwork(NetworkGraphMixin, SubNetworkPowerFlowMixin):
         """
         return self.c.stores.static
 
-    @deprecated_in_next_major(details="Use `self.components` instead.")
+    @deprecated(
+        deprecated_in="1.0.0",
+        removed_in="2.0.0",
+        details="Use `self.components` instead.",
+    )
     # Deprecate: Use `self.iterate_components` instead
     def iterate_components(
         self, components: Collection[str] | None = None, skip_empty: bool = True